+++ packages/cad-viewer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mlightcad/cad-viewer
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/cad-simple-viewer@1.1.5
+
 ## 1.1.4
 
 ### Patch Changes