+++ packages/three-renderer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/three-renderer
 
+## 1.1.5
+
+### Patch Changes
+
+- feat: fix bugs on rendering insert (block reference) entity which contains entities in different layers
+
 ## 1.0.20
 
 ### Patch Changes